pyproject.toml
[tool.poetry]
name = "feedgen_hasname"
version = "0.2020.0628.0"
description = "feedgen.hasname.com"

license = "MIT"

authors = [
    "Gea-Suan Lin <gslin@gslin.com>"
]

readme = "README.md"

repository = "https://github.com/hasname/feedgen.git"
homepage = "https://github.com/hasname/feedgen"

keywords = ["feed"]

[tool.poetry.dependencies]
base36 = "*"
cssselect = "*"
django = "*"
django-service-objects = "*"
feedgen = "*"
lxml = "*"
pip = "*"
pysocks = "*"
python = "^3.8"
python-dateutil = "*"
python-dotenv = "*"
requests = "*"
requests_futures = "*"
sentry-sdk = "*"
uwsgi = "*"

[tool.poetry.dev-dependencies]
coverage = "*"
requests-mock = "*"
